[REFAC] Rename screen directory to page
